+++ mantisshrimp/core/__init__.py
@@ -1,4 +1,5 @@
 from mantisshrimp.core.category import *
 from mantisshrimp.core.id_map import *
+from mantisshrimp.core.class_map import *
 from mantisshrimp.core.mask import *
 from mantisshrimp.core.bbox import *